autenticazione, middleware e mail

+++ routes/web.php
@@ -1,39 +1,48 @@
 <?php
 
+use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BlogController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\UserController;
+use App\Http\Middleware\IsAdmin;
 
-
-
-//route principale della dashboard
 Route::get('/', function () {
     return view('welcome');
 });
 
-//route per la pagina dei blog
-// Route::get('/blogs', function () {
-//     return view('blogs.index');
-// });
+Route::get('/dashboard', function () {
+    return view('dashboard');
+})->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::get('/blogs', [BlogController::class, 'index']);
+Route::middleware('auth')->group(function () {
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-Route::get('/test-post', [BlogController::class, 'getRequest']);
-Route::post('/test-post', [BlogController::class, 'postRequest']);
+    Route::get('/blogs', [BlogController::class, 'index']);
 
-Route::get('/test-par/{id}', [BlogController::class, 'getPar']);
+    Route::get('/test-post', [BlogController::class, 'getRequest']);
+    Route::post('/test-post', [BlogController::class, 'postRequest']);
 
-Route::get('getfile', [BlogController::class, 'getFile']);
+    Route::get('/test-par/{id}', [BlogController::class, 'getPar']);
 
-Route::get('posts', [PostController::class, 'index']);
+    Route::get('getfile', [BlogController::class, 'getFile']);
 
-Route::get('create-post', [PostController::class, 'create']);
-Route::post('create-post', [PostController::class, 'savePost']);
+    Route::get('posts', [PostController::class, 'index'])->middleware(IsAdmin::class);
 
-Route::get('edit-post/{id}', [PostController::class, 'edit']);
-Route::post('update-post/{id}', [PostController::class, 'updatePost']);
+    Route::get('create-post', [PostController::class, 'create'])->middleware(IsAdmin::class);
+    Route::post('create-post', [PostController::class, 'savePost'])->middleware(IsAdmin::class);
 
-Route::delete('delete-post/{id}', [PostController::class, 'delete']);
+    Route::get('edit-post/{id}', [PostController::class, 'edit'])->middleware(IsAdmin::class);
+    Route::post('update-post/{id}', [PostController::class, 'updatePost'])->middleware(IsAdmin::class);
 
-Route::get('user/{id}', [UserController::class, 'userProfile']);
+    Route::delete('delete-post/{id}', [PostController::class, 'delete'])->middleware(IsAdmin::class);
+
+    Route::get('user/{id}', [UserController::class, 'userProfile']);
+
+});
+
+
+
+require __DIR__.'/auth.php';
